The push was rejected because another session had advanced the same branch by 16
commits, including 03a2ad44e "integrate Tensors 0.121.0 (#1946)" — a deliberate
integration of the same package bump I had hand-resolved locally against master.

Reassuringly, both arrived at the SAME version matrix (Tensors 0.121.0, OneDNN
0.120.6, OpenBLAS 0.121.0, CLBlast 0.120.6), so the max-of-each reasoning in the
previous merge held up against an independent integration.

The only conflict was comment prose. Both notes are kept because they document
different bumps and do not contradict: mine explains why OneDNN/CLBlast stay at
0.120.6 rather than following master back to 0.119.0 (#906 saved-state pinning,
#907 LSTM GPU kernels), theirs explains the 0.121.0 move (#925, full-span autodiff
views returned to the pool while the graph still owns them).

Verified: dotnet restore exits 0 with zero errors.

+++ src/OnlineLearning/OnlineLearningModelBase.cs
@@ -183,7 +183,13 @@ public abstract class OnlineLearningModelBase<T> : IOnlineLearningModel<T>, IMod
             LearningRateSchedule.Constant => InitialLearningRate,
             LearningRateSchedule.InverseScaling => InitialLearningRate / (1.0 + 0.0001 * SampleCount),
             LearningRateSchedule.Exponential => InitialLearningRate * Math.Pow(0.9999, SampleCount),
-            LearningRateSchedule.StepDecay => InitialLearningRate * Math.Pow(0.5, SampleCount / 10000),
+            // The truncation here is DELIBERATE and is what makes this a step decay:
+            // the rate halves once per 10,000 samples and is flat between steps.
+            // Written as an explicit integer division so it reads as intent rather
+            // than as the accidental int/int that a reviewer (and CodeQL) sees. Using
+            // 10000.0 would silently turn this into a smooth exponential decay and
+            // change every schedule that selects StepDecay.
+            LearningRateSchedule.StepDecay => InitialLearningRate * Math.Pow(0.5, (double)(SampleCount / 10000)),
             _ => InitialLearningRate
         };
